+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - [UT: add metadata-transfer unit test](https://github.com/Tencent/spring-cloud-tencent/pull/335)
 - [feat:Add GitHub action of codecov.yml and update junit of metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/344)
 - [Feature: optimize static metadata manager](https://github.com/Tencent/spring-cloud-tencent/pull/346)
+- [fix: shutdown thread pool before the container closes](https://github.com/Tencent/spring-cloud-tencent/pull/354)